refactor(selecao): extrai records de request do ProcessoSeletivoController
