+++ src/selecao/Unifesspa.UniPlus.Selecao.API/Contracts/Requests/AlterarMotivoRetificacaoRequest.cs
@@ -1,0 +1,8 @@
+namespace Unifesspa.UniPlus.Selecao.API.Contracts.Requests;
+
+using Controllers;
+
+/// <summary>
+/// Corpo de <see cref="ProcessoSeletivoController.AlterarMotivoRetificacao"/>.
+/// </summary>
+public sealed record AlterarMotivoRetificacaoRequest(string Motivo);
