+++ src/selecao/Unifesspa.UniPlus.Selecao.API/Controllers/ProcessoSeletivoController.cs
@@ -6,6 +6,8 @@ using Application.Commands.ProcessosSeletivos;
 using Application.DTOs;
 using Application.Queries.ProcessosSeletivos;
 
+using Contracts.Requests;
+
 using Http;
 
 using Microsoft.AspNetCore.Authorization;
@@ -943,125 +945,3 @@ public sealed class ProcessoSeletivoController : ControllerBase
         return resultado.IsSuccess ? Ok(resultado.Value) : resultado.ToActionResult(_mapper);
     }
 }
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.Publicar"/> — omite
-/// <c>ProcessoSeletivoId</c> (vem da rota).
-/// </summary>
-public sealed record PublicarProcessoSeletivoRequest(
-    string? Numero,
-    DateOnly PeriodoInscricaoInicio,
-    DateOnly PeriodoInscricaoFim,
-    Guid DocumentoEditalId,
-    DadosDoAtoRequest Ato);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.Retificar"/> — carrega só os
-/// dados próprios da retificação. Omite <c>ProcessoSeletivoId</c> (vem da rota)
-/// e não recebe id de Edital: o Edital sucedido é o vigente, resolvido no
-/// servidor (a retificação endereça o agregado, não uma entidade interna).
-/// </summary>
-public sealed record RetificarProcessoSeletivoRequest(
-    string Motivo,
-    string? Numero,
-    DateOnly PeriodoInscricaoInicio,
-    DateOnly PeriodoInscricaoFim,
-    Guid DocumentoEditalId,
-    DadosDoAtoRequest Ato);
-
-/// <summary>
-/// Dados que o DOCUMENTO declara sobre si — órgão publicador, série, ano, data de
-/// publicação, quem assina e o tipo do ato no catálogo de Publicações. Nenhum deles é
-/// derivado pelo sistema: a data documental não é o relógio, o assinante não é o usuário
-/// autenticado, e o tipo não se infere (ADR-0103 — um aviso pode retificar um edital).
-/// </summary>
-public sealed record DadosDoAtoRequest(
-    string Orgao,
-    string Serie,
-    int Ano,
-    DateOnly DataPublicacao,
-    string Assinante,
-    string TipoAtoCodigo);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.AbrirRetificacao"/> — só o motivo. A
-/// versão que a sessão retifica <b>não</b> é informada pelo cliente: o servidor a infere do
-/// topo da cadeia (ADR-0101), sob o mesmo lock que abre o rascunho.
-/// </summary>
-public sealed record AbrirRetificacaoRequest(string Motivo);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.AlterarMotivoRetificacao"/>.
-/// </summary>
-public sealed record AlterarMotivoRetificacaoRequest(string Motivo);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.FecharRetificacao"/> — o do atalho atômico
-/// <b>menos o motivo</b>, que já foi declarado na abertura e vive no rascunho.
-/// </summary>
-public sealed record FecharRetificacaoRequest(
-    string? Numero,
-    DateOnly PeriodoInscricaoInicio,
-    DateOnly PeriodoInscricaoFim,
-    Guid DocumentoEditalId,
-    DadosDoAtoRequest Ato);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.DefinirOfertaAtendimento"/>
-/// — omite <c>ProcessoSeletivoId</c> (vem da rota).
-/// </summary>
-public sealed record DefinirOfertaAtendimentoRequest(
-    IReadOnlyList<Guid> CondicaoIds,
-    IReadOnlyList<Guid> RecursoIds,
-    IReadOnlyList<Guid> TipoDeficienciaIds);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.DefinirBonusRegional"/> —
-/// omite <c>ProcessoSeletivoId</c> (vem da rota).
-/// </summary>
-public sealed record DefinirBonusRegionalRequest(
-    string? RegraCodigo,
-    string? RegraVersao,
-    decimal? Fator,
-    decimal? Teto,
-    string? MunicipioConvenio,
-    string? BaseLegal);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.DefinirReferenciaTemporalFatos"/> —
-/// omite <c>ProcessoSeletivoId</c> (vem da rota).
-/// </summary>
-/// <remarks>
-/// Contrato por variante de <c>Tipo</c> (Story #554, PR #896 — issue #892; tokens em
-/// <see cref="Domain.Enums.ReferenciaTipoCodigo"/>), tudo-ou-nada por linha (N-I01):
-/// <list type="table">
-/// <listheader><term>Tipo</term><description>Data</description><description>FaseId</description></listheader>
-/// <item><term><see langword="null"/> (remove a referência)</term><description>proibido</description><description>proibido</description></item>
-/// <item><term><c>FIM_INSCRICAO</c></term><description>proibido</description><description>proibido</description></item>
-/// <item><term><c>INICIO_FASE</c> / <c>FIM_FASE</c></term><description>proibido</description><description>obrigatório</description></item>
-/// <item><term><c>DATA_ESPECIFICA</c></term><description>obrigatório</description><description>proibido</description></item>
-/// </list>
-/// Violação de qualquer linha recusa com <c>ReferenciaTemporalFatos.DataIncoerenteComTipo</c>
-/// ou <c>ReferenciaTemporalFatos.FaseIncoerenteComTipo</c> (422) — sem fallback silencioso
-/// (ADR-0111:235-236). <c>FaseId</c>, quando presente, precisa pertencer ao cronograma do
-/// MESMO processo (<c>ReferenciaTemporalFatos.FaseNaoPertenceAoProcesso</c>).
-/// </remarks>
-public sealed record DefinirReferenciaTemporalFatosRequest(
-    string? Tipo,
-    DateOnly? Data,
-    Guid? FaseId);
-
-/// <summary>
-/// Corpo de <see cref="ProcessoSeletivoController.DefinirClassificacao"/> —
-/// omite <c>ProcessoSeletivoId</c> (vem da rota).
-/// </summary>
-public sealed record DefinirClassificacaoRequest(
-    string RegraCalculoCodigo,
-    string RegraCalculoVersao,
-    string? RegraArredondamentoCodigo,
-    string? RegraArredondamentoVersao,
-    int? CasasArredondamento,
-    string RegraOrdemAlocacaoCodigo,
-    string RegraOrdemAlocacaoVersao,
-    int NOpcoesAlocacao,
-    IReadOnlyList<RegraEliminacaoInput> RegrasEliminacao);
